+++ src/Catalog/CatalogService.php
@@ -126,6 +126,31 @@ final class CatalogService extends AbstractService
         );
     }
 
+    /** @return AttributeVariant[] */
+    public function getAttributeVariants(int $attributeId): array
+    {
+        try {
+            $variantsData = $this->client->get("catalog/attributes/$attributeId/variants");
+        } catch (ClientException $e) {
+            if ($e->getResponse()->getStatusCode() === 404) {
+                throw new NotFound("Attribute #$attributeId not found", $e);
+            }
+            throw $e;
+        }
+
+        return array_map(function (array $variantData): AttributeVariant {
+            $image = (!empty($variantData['image'])) ? new Image($variantData['image']) : null;
+
+            return new AttributeVariant(
+                $variantData['id'],
+                $variantData['attributeId'],
+                $variantData['name'],
+                $variantData['slug'],
+                $image
+            );
+        }, $variantsData);
+    }
+
     /**
      * Report a suspicious product to the marketplace administrator.
      *