+++ tests/Catalog/CatalogServiceTest.php
@@ -9,6 +9,7 @@ namespace Wizaplace\Tests\Catalog;
 
 use GuzzleHttp\Psr7\Response;
 use Wizaplace\Catalog\AttributeType;
+use Wizaplace\Catalog\AttributeVariant;
 use Wizaplace\Catalog\CatalogService;
 use Wizaplace\Catalog\Declination;
 use Wizaplace\Catalog\Option;
@@ -286,6 +287,25 @@ final class CatalogServiceTest extends ApiTestCase
     {
         $this->expectException(NotFound::class);
         $this->buildCatalogService()->getAttributeVariant(404);
+    }
+
+    public function testGetAttributeVariants()
+    {
+        $variants = $this->buildCatalogService()->getAttributeVariants(1);
+
+        $expectedVariants = [
+            new AttributeVariant(1, 1, 'Bleu', 'bleu', null),
+            new AttributeVariant(2, 1, 'Blanc', 'blanc', null),
+            new AttributeVariant(3, 1, 'Rouge', 'rouge', null),
+        ];
+
+        $this->assertEquals($expectedVariants, $variants);
+    }
+
+    public function testGetNonExistingAttributeVariants()
+    {
+        $this->expectException(NotFound::class);
+        $this->buildCatalogService()->getAttributeVariants(404);
     }
 
     public function testGetProductWithOptions()